sample/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
openai==1.52.0
anthropic==0.37.0
langchain==0.3.4
langchain-openai==0.2.3
langchain-community==0.3.3
chromadb==0.5.12
tiktoken==0.8.0
httpx==0.27.2
pydantic==2.9.0
python-jose==3.3.0
cryptography==43.0.0
bcrypt==4.2.0
redis==5.1.0
sqlalchemy==2.0.35
boto3==1.35.30
structlog==24.4.0
sentry-sdk==2.14.0
requests==2.32.3
strands-agents==0.1.5
strands-agents-tools==0.1.2
guardrails-ai==0.5.0
mcp==1.1.0
